// Backend/src/matches/enums/match-event-type.enum.ts
export enum MatchEventType {
  GOAL = 'goal',
  YELLOW_CARD = 'yellowCard',
  RED_CARD = 'redCard',
  BLUE_CARD = 'blueCard',
  CARD = 'card',
  START_FIRST_HALF = 'start_first_half',
  END_FIRST_HALF = 'end_first_half',
  START_SECOND_HALF = 'start_second_half',
  END_SECOND_HALF = 'end_second_half',
}